+++ streamlabsio/__init__.py
@@ -1,1 +1,7 @@
-from .client import connect
+from loguru import logger
+
+from .client import request_client_object as connect
+
+__ALL__ = ['connect']
+
+logger.disable('streamlabsio')